+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Mention Docker and the NMDC EDGE implementation in the `README.md`.
 
 ## [1.3.3] - 2023-01-05
 ### Fixed